+++ docs/pages/9 - Contrib Modules.md
@@ -566,105 +566,6 @@ object app extends ScalaModule with RouterModule {
 }
 ```
 
-## Release
-
-This plugin helps set release versions.
-
-### Quickstart
-```scala
-import $ivy.`com.lihaoyi::mill-contrib-release:$MILL_VERSION`
-import mill.contrib.release.ReleaseModule
-
-object release extends ReleaseModule
-
-object mymodule extends PublishModule {
-  def publishVersion = release.currentVersion().toString
-  ...
-}
-```
-
-Then to release a new version and update version to next:
-```
-$ mill release.setReleaseVersion
-$ mill mymodule.publish
-$ mill release.setNextVersion --bump minor
-```
-
-### Configure the version file
-This module stores the version in a version file in the repository root.
-The filename being used is set by `ReleaseModule.versionFile` and defaults to `version`, but this can
-be overridden if you prefer something else than the default:
-
-```scala
-object release extends ReleaseModule {
-  def versionFile = ".version"
-}
-```
-
-### Set release version
-The `setReleaseVersion` target removes the `-SNAPSHOT` identifier from the version.
-Then it commits and adds a tag.
-
-#### Example
-Your version file contains `0.1.0-SNAPSHOT`. In your terminal you do the following:
-```bash
-$ mill release.setReleaseVersion
-```
-
-First it will update the version file to contain `0.1.0`.
-Then it will execute:
-```bash
-git commit -am "Setting release version to 0.1.0"
-git tag 0.1.0
-```
-
-### Set next version
-The `setNextVersion` target bumps the version and sets this as the new snapshot version.
-Then it commits **and pushes to origin/master with tags**.
-
-#### Parameters
-
-##### --bump (major | minor | patch)
-Sets what segment of the version to bump.
-
-For a version number `1.2.3` in the version file:
-
-`--bump major` will set it to `2.0.0`
-
-`--bump minor` will set it to `1.3.0`
-
-`--bump patch` will set it to `1.2.4`
-
-#### Example
-Your version file contains `0.1.0`. In your terminal you do the following:
-```bash
-$ mill release.setNextVersion --bump minor
-```
-
-First it will update the version file to contain `0.2.0-SNAPSHOT`.
-Then it will execute:
-```bash
-git commit -am "Setting next version to 0.2.0-SNAPSHOT"
-git push origin master --tags
-```
-
-### Output version numbers
-If you need to output the version numbers (for example for other CI tools you might use), you can use the following commands:
-```bash
-# Show the current version from the version file.
-$ mill show release.currentVersion
-```
-
-```bash
-# Show the version that would be used as release version.
-$ mill show release.releaseVersion
-```
-
-```bash
-# Show the version that would be used as next version with the given --bump argument.
-$ mill show release.nextVersion --bump minor
-```
-
 ## ScalaPB
 
 This module allows [ScalaPB](https://scalapb.github.io) to be used in Mill builds. ScalaPB is a [Protocol Buffers](https://developers.google.com/protocol-buffers/) compiler plugin that generates Scala case classes, encoders and decoders for protobuf messages.
@@ -943,3 +844,118 @@ These imports will always be added to every template.  You don't need to list th
 
 ### Example
 There's an [example project](https://github.com/lihaoyi/cask/tree/master/example/twirl)
+
+## Version file
+
+This plugin provides helpers for updating a version file and committing the changes to git.
+
+**Note: You can still make manual changes to the version file in-between execution of the targets provided by the module.**
+**Each target operates on the version file as is at the time of execution.**
+
+### Quickstart
+```scala
+import $ivy.`com.lihaoyi::mill-contrib-release:$MILL_VERSION`
+import mill.contrib.release.ReleaseModule
+
+object release extends ReleaseModule
+```
+
+Then to commit a release version and then commit next snapshot version:
+```bash
+$ mill release.setReleaseVersion
+$ mill release.setNextVersion --bump minor
+```
+
+You can also make manual changes in-between:
+```bash
+$ mill release.setReleaseVersion
+$ echo 0.1.0 > version
+# Will now set the version to 0.2.0-SNAPSHOT
+$ mill release.setNextVersion --bump minor
+```
+
+If you want to use the version file for publishing, you can do it like this:
+```scala
+import $ivy.`com.lihaoyi::mill-contrib-versionfile:$MILL_VERSION`
+import mill.contrib.versionfile.VersionFileModule
+
+object release extends VersionFileModule
+
+object mymodule extends PublishModule {
+  def publishVersion = release.currentVersion().toString
+  ...
+}
+```
+
+### Configure the version file
+This module reads and updates a version file. The default is the file `version` at the project root.
+
+If you prefer something else than the default, you can override:
+* `VersionFileModule.versionFileName` which defaults to `version`
+* `VersionFileModule.versionFileDirectory` which defaults to `os.pwd`
+* `VersionFileModule.versionFilePath` which defaults to `versionFileDirectory / versionFileName`
+
+### Set release version
+The `setReleaseVersion` target removes the `-SNAPSHOT` identifier from the version.
+Then it overwrites the previous content in the version file, commits the changes and adds a tag.
+
+#### Example
+Your version file contains `0.1.0-SNAPSHOT`. In your terminal you do the following:
+```bash
+$ mill release.setReleaseVersion
+```
+
+First it will update the version file to contain `0.1.0`.
+Then it will execute:
+```bash
+git commit -am "Setting release version to 0.1.0"
+git tag 0.1.0
+```
+
+### Set next version
+The `setNextVersion` target bumps the version and sets this as the new snapshot version.
+Then it overwrites the previous content in the version file, commits the changes
+**and pushes to origin/master with tags**.
+
+#### Parameters
+
+##### --bump (major | minor | patch)
+Sets what segment of the version to bump.
+
+For a version number `1.2.3` in the version file:
+
+`--bump major` will set it to `2.0.0`
+
+`--bump minor` will set it to `1.3.0`
+
+`--bump patch` will set it to `1.2.4`
+
+#### Example
+Your version file contains `0.1.0`. In your terminal you do the following:
+```bash
+$ mill release.setNextVersion --bump minor
+```
+
+First it will update the version file to contain `0.2.0-SNAPSHOT`.
+Then it will execute:
+```bash
+git commit -am "Setting next version to 0.2.0-SNAPSHOT"
+git push origin master --tags
+```
+
+### Output version numbers
+If you need to output the version numbers (for example for other CI tools you might use), you can use the following commands:
+```bash
+# Show the current version from the version file.
+$ mill show release.currentVersion
+```
+
+```bash
+# Show the version that would be used as release version.
+$ mill show release.releaseVersion
+```
+
+```bash
+# Show the version that would be used as next version with the given --bump argument.
+$ mill show release.nextVersion --bump minor
+```
